1微信退款及通知完善，2加入数据库日志logback  3加解密工具
1退款完善，2加入数据库日志logback 3加解密工具

+++ gpmall-shopping/src/main/java/com/gpmall/shopping/controller/OrderController.java
@@ -87,7 +87,7 @@ public class OrderController {
         OrderDetailResponse response=orderQueryService.orderDetail(request);
         if(response.getCode().equals(OrderRetCode.SUCCESS.getCode())){
             OrderDetail orderDetail=new OrderDetail();
-            orderDetail.setOrderTotal(BigDecimal.valueOf(response.getPayment()));
+            orderDetail.setOrderTotal(response.getPayment());
             orderDetail.setUserId(response.getUserId());
             orderDetail.setUserName(response.getBuyerNick());
             orderDetail.setGoodsList(response.getOrderItemDto());
